feat(subconscious): phase 2 — extract the memory profile (behavior-identical)
Move the old engine's stages 1–3 into profiles/memory.rs as MemoryProfile and
switch the live path to a memory SubconsciousInstance; delete the monolithic
engine.rs. Pure extraction — no behavior change.

- profiles/memory.rs: MemoryProfile (observe = memory_diff, prepare_context =
  context_scout with root parent, reflect = slim decision agent, commit =
  re-checkpoint baseline, origin = taint) + render_world_diff / tick_origin_source
  + `memory_instance(config)` constructor.
- The tiny.place orchestration_review still piggybacks on the memory tick — now
  a clearly-marked `phase-3 removes this` shim at the top of MemoryProfile::observe
  (phase 3 gives it its own instance).
- global.rs / subconscious_cli.rs build via `memory_instance`; mod.rs keeps
  `SubconsciousEngine` as a type alias for SubconsciousInstance.
- Delete engine.rs + engine_tests.rs; the tick-origin and world-diff-render tests
  move to profiles/memory_tests.rs, the provider/rate-cap pure tests to
  provider_tests.rs, and the rate-cap state-machine test to instance_tests.rs.

All 124 subconscious lib tests pass.
